change GetFilePath param from int to IntPtr Previously the GetFilePath method of NotepadPPGateway was unusable in 64-bit Notepad++ because it accepted an int parameter, but the buffer id (the IdFrom parameter of the NPPN_FILEBEFORECLOSE notification) could be 64-bit.
Also added a new plugin command to Demo.cs illustrating how
tracking the NPPN_FILEBEFORECLOSE notification allows
the plugin to keep a list of all the filenames that were
closed this session.

+++ Visual Studio Project Template C#/PluginInfrastructure/NotepadPPGateway.cs
@@ -16,7 +16,7 @@ namespace Kbg.NppPluginNET.PluginInfrastructure
 		string GetNppPath();
 		string GetPluginConfigPath();
 		string GetCurrentFilePath();
-		unsafe string GetFilePath(int bufferId);
+		unsafe string GetFilePath(IntPtr bufferId);
 		void SetCurrentLanguage(LangType language);
 		bool OpenFile(string path);
 	}
@@ -108,7 +108,7 @@ namespace Kbg.NppPluginNET.PluginInfrastructure
 		/// <summary>
 		/// Gets the path of the current document.
 		/// </summary>
-		public unsafe string GetFilePath(int bufferId)
+		public unsafe string GetFilePath(IntPtr bufferId)
 		{
 			var path = new StringBuilder(2000);
 			Win32.SendMessage(PluginBase.nppData._nppHandle, (uint) NppMsg.NPPM_GETFULLPATHFROMBUFFERID, bufferId, path);
